add transformer global options

+++ src/configs/microsoft/option_values.rs
@@ -7,7 +7,7 @@ use crate::validators::validate_string_optional;
 #[derive(Debug, Deserialize)]
 #[serde(rename = "OptionValue", rename_all = "PascalCase")]
 pub struct MicrosoftOptionValue {
-    pub option_id: u32,
+    pub option_id: u8,
     pub value: Option<Vec<String>>,
     #[serde(deserialize_with = "validate_string_optional")]
     pub vendor_class: Option<String>,
